project/requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
branca==0.4.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
cycler==0.10.0
Fiona==1.8.18
flake8==3.8.4
folium==0.11.0
geojson==2.5.0
geopandas==0.8.1
idna==2.10
Jinja2==3.0.0a1
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==2.0.0a1
matplotlib==3.3.3
mccabe==0.6.1
mpmath==1.1.0
munch==2.5.1.dev12
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.4
pathspec==0.8.1
patsy==0.5.1
Pillow==8.0.1
powerlaw==1.4.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==3.0.0b1
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scikit-mobility==1.1.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
statsmodels==0.12.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.52.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
